+++ packages/win_toast/windows/notification_manager.cc
@@ -32,3 +32,4 @@ bool NotificationManager::HasIdentity() {
 
   return _hasIdentity;
 }
+